% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{royaltweets}
\alias{royaltweets}
\title{All royal tweets .}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75000 rows and 18 columns.}
\usage{
royaltweets
}
\description{
All royal tweets .
}
\keyword{datasets}
